refactor(macos): split CEFRenderer cookies into extension file
Extract cookie read/write/delete operations, the CDP-then-JS fallback
machinery, and the cookieInjectionScript builder out of CEFRenderer.swift
into CEFRenderer+Cookies.swift. CEFRenderer.swift now stays under the
500-line limit. CEFHostView, CookieContinuationState, bridge, containerView,
pendingCookies, and pendingDeleteAllCookies are now internal so the
extension in a sibling file can reach them. Pure code motion.
